Clean up flash file and start to deprecate Grayskull
- Create wormhole.py for symmetry with blackhole.py, handles the legacy
SPI format
- Split out functions that create FlashWrite data for WH and BH into
their own files (wormhole.py and blackhole.py respectively).
- Skip flashing Grayskull/any other chips if they are detected.
- Format edited files with black
